+++ pydec/bias_decomposition.py
@@ -3,8 +3,9 @@ import functools
 import inspect
 import torch
 from torch import Tensor
+from torch.autograd.grad_mode import _DecoratorContextManager
 
-from typing import Dict, Union, Any, Callable, ContextManager, TYPE_CHECKING
+from typing import Dict, Union, Any, Callable, TYPE_CHECKING
 
 if TYPE_CHECKING:
     from .composition import Composition
@@ -35,6 +36,7 @@ class _BiasDecompositionState:
 
 def register_bias_decomposition_func(name):
     """
+    TODO: need update
     New bias_decomposition_func can be added with the :func:`register_bias_decomposition_func`
     function decorator.
 
@@ -104,7 +106,7 @@ def get_bias_decomposition_func() -> Callable[..., Composition]:
         return current_bias_decomposition_func
 
 
-class using_bias_decomposition_func(ContextManager):
+class using_bias_decomposition_func(_DecoratorContextManager):
     def __init__(self, name: str) -> None:
         if name not in _BIAS_DECOMPOSITION_FUNC_REGISTRY:
             raise ValueError(
@@ -120,8 +122,11 @@ class using_bias_decomposition_func(ContextManager):
     def __exit__(self, exc_type: Any, exc_value: Any, traceback: Any) -> None:
         _BiasDecompositionState.bias_decomposition_name = self.prev
 
+    def clone(self):
+        return self.__class__(self.using_name)
 
-class no_bias_decomposition(ContextManager):
+
+class no_bias_decomposition(_DecoratorContextManager):
     def __init__(
         self,
     ) -> None:
@@ -146,7 +151,7 @@ def get_bias_decomposition_args() -> Dict[str, Any]:
     return _BiasDecompositionState.bias_decomposition_args
 
 
-class using_bias_decomposition_args(ContextManager):
+class using_bias_decomposition_args(_DecoratorContextManager):
     def __init__(self, update=True, **kwargs) -> None:
         self.update = update
         self.prev = None
@@ -162,9 +167,14 @@ class using_bias_decomposition_args(ContextManager):
     def __exit__(self, exc_type: Any, exc_value: Any, traceback: Any) -> None:
         _BiasDecompositionState.bias_decomposition_args = self.prev
 
+    def clone(self):
+        return self.__class__(self.update, **self.using_args)
+
 
 @register_bias_decomposition_func("none")
-def _none_decomposition(bias: Union[Number, Tensor], context: Composition):
+def _none_decomposition(
+    bias: Union[Number, Tensor], context: Composition
+) -> Composition:
     r"""
     Default decomposition with no_bias_decomposition. Just add the bias to residual.
     """
@@ -251,8 +261,8 @@ def sign_decomposition(
     return out
 
 
-@register_bias_decomposition_func("sign_decomposition_value_threshold")
-def sign_decomposition_value_threshold(
+@register_bias_decomposition_func("sign_decomposition_threshold")
+def sign_decomposition_threshold(
     bias: Union[Number, Tensor],
     context: Composition,
     *,
@@ -273,8 +283,8 @@ def sign_decomposition_value_threshold(
     return out
 
 
-@register_bias_decomposition_func("hybrid_decomposition_value_threshold")
-def hybrid_decomposition_value_threshold(
+@register_bias_decomposition_func("hybrid_decomposition_threshold")
+def hybrid_decomposition_threshold(
     bias: Union[Number, Tensor],
     context: Composition,
     *,